did refactoring of UI file placement to avoid conflicts later
