+++ app/src/androidTest/java/at/aau/serg/websocketbrokerdemo/StartScreenTest.kt
@@ -3,6 +3,7 @@ package at.aau.serg.websocketbrokerdemo
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.*
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import at.aau.serg.websocketbrokerdemo.ui.StartScreen
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
